+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 
 (unreleased)
 ------------
+- Feat: Add date to the email message. [Stavros Korokithakis]
 - Feat: Add prompt customization. [Stavros Korokithakis]
 - Initial commit. [Stavros Korokithakis]